Move schema validation from Configuration.ts to /schema directory.
