+++ packages/realm/src/schema.ts
@@ -20,3 +20,4 @@ export * from "./schema/from-binding";
 export * from "./schema/to-binding";
 export * from "./schema/normalize";
 export * from "./schema/types";
+export * from "./schema/validate";
